+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Upcoming Release
 
 ## Changed
-- Do not try to install a programm with more than 65,535 instructions
+- Do not try to install a programm with more than 4096 instructions
 
 # v0.4.0